Refactor to base class
